% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useTemplate.R
\name{previewTemplate}
\alias{previewTemplate}
\title{Preview basic boilerplates of the RinteRface HTML template collection}
\usage{
previewTemplate(lib = c("shinydashboardPlus", "bs4Dash", "argonDash",
  "tablerDash"))
}
\arguments{
\item{lib}{Choose a flavor among \code{c("shinydashboardPlus", "bs4Dash", "argonDash", "tablerDash")}.}
}
\description{
Preview basic boilerplates of the RinteRface HTML template collection
}
\examples{
if (interactive()) {
 previewTemplate("bs4Dash")
}

}
