% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useTemplate.R
\name{usePage}
\alias{usePage}
\title{Select the relevant dashboard page function}
\usage{
usePage(lib)
}
\arguments{
\item{lib}{Choose a flavor among \code{c("shinydashboardPlus", "bs4Dash", "argonDash", "tablerDash")}.}
}
\description{
Depending on the selected library, choose the suitable dashboard page function
}
