% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useTemplate.R
\name{useTemplate}
\alias{useTemplate}
\title{Import RinteRface boilerplates in the chosen directory}
\usage{
useTemplate(path, lib)
}
\arguments{
\item{path}{Name of the folder to copy the boilerplates in.}

\item{lib}{Choose a flavor among \code{c("shinydashboardPlus", "bs4Dash", "argonDash", "tablerDash")}.}
}
\description{
Import RinteRface boilerplates in the chosen directory
}
